+++ inc/cache.h
@@ -17,6 +17,18 @@
 
 extern std::array<O3_CPU*, NUM_CPUS> ooo_cpu;
 
+// CACHE BLOCK
+class BLOCK
+{
+public:
+  bool valid = false, prefetch = false, dirty = false;
+
+  uint64_t address = 0, v_address = 0, tag = 0, data = 0, ip = 0, cpu = 0, instr_id = 0;
+
+  // replacement state
+  uint32_t lru = std::numeric_limits<uint32_t>::max() >> 1;
+};
+
 class CACHE : public champsim::operable, public MemoryRequestConsumer, public MemoryRequestProducer
 {
 public:
